Change Unicode property abbrev to upcoming official
Unicode 12.0 used a new property file that was not from the Unicode
Character Database.  It only had a long property name.  I incorporated
it into our data, and rather than use the very long name all the time, I
created my own short name, since there was no official one.

Now, the upcoming 13.0 has moved the file to the UCD, and come up with a
short name that differs from the one I had.  This commit converts to use
Unicode's name.  This property is not exposed to user or XS space, so
there is no user impact.